requirements.txt
telethon
google-generativeai
MetaTrader5
python-dotenv
pydantic
pydantic-settings